backend/requirements.txt
python-dotenv~=1.0.1
Flask~=3.1.0
bcrypt~=4.2.1
psycopg2~=2.9.10
PyJWT~=2.10.1
Flask-Cors~=5.0.0